refactor(refs T30969): move DpStickyElement to components root
